requirements.txt
aiogram==3.7.0
PyYAML==6.0.1